requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
scikit-learn==1.3.1
tqdm==4.66.1
python-dotenv==1.0.0
networkx>=2.5,<3.0
node2vec==0.4.6
matplotlib==3.8.0
pyvis==0.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.22.0
